# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Crowdfunding Toolbox",
        "description": "",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "stefan.ondas@proup.sk"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/backoffice/login": {
            "post": {
                "tags": [
                    "BACKOFFICE USER"
                ],
                "summary": "Login backoffice user",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "username / email": "test / test@test.com",
                                "password": "test123"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully logged in!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/backoffice/register": {
            "post": {
                "tags": [
                    "BACKOFFICE USER"
                ],
                "summary": "Create backoffice user",
                "description": "Only for users with admin role. They can create new users.",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "email": "test@test.com",
                                "username": "test",
                                "firstName": "John",
                                "lastName": "Doe",
                                "password": "test123",
                                "confirmationPassword": "test123"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created user!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/backoffice/crowdfunding-settings": {
            "put": {
                "tags": [
                    "GENERAL SETTINGS"
                ],
                "summary": "Crowdfunding Settings",
                "description": "Only for authenticated users.",
                "operationId": "crowdFundingSettings",
                "requestBody": {
                    "description": "After created campaign, it will be activated.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "settings": {
                                        "type": "json"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "type": "general",
                                "settings": {
                                    "color": {
                                        "primary": "#000000",
                                        "secondary": "#ff0000"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created campaign!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/backoffice/reset-password": {
            "post": {
                "tags": [
                    "BACKOFFICE USER"
                ],
                "summary": "Reset password for backoffice user",
                "operationId": "App\\Http\\Controllers\\BackOfficeAPI\\UsersController::resetPassword",
                "requestBody": {
                    "description": "Property role can be 1 or 2. 1 for Admin user, 2 for classic user.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "email": "test@test.com"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Email with new password was successfully sent.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Email is not exist"
                    }
                }
            }
        },
...
    }
}